[Refactoring] GUI Modulize
