+++ gui/__init__.py
@@ -1,0 +1,6 @@
+# utils/__init__.py
+# 이 패키지는 코드 추출, 실행, 처리 및 파이프라인 기능을 포함합니다.
+
+from .ai_panel import AiPanel
+from .code_panel import CodePanel
+from .side_panel import SidePanel
